First action! :D
A pretty naive solution to a leetcode challenge.
Just started to learn python, and as time shows the best way is to engage in it directly. For a moment I lack language's features, so I decided to start with things that are common everywhere :) LOOPS!!!

Will definitely work on time complexity here.