Add files via upload
The idea behind this solution is to update the end pointer by pointing it at a target_index when the value < 0. In other cases just advance the start pointer until it reaches the end pointer. It needs a bit of refactoring, since the last if statement in the bin() is clumsy